requirements.txt
# requirements.txt

PyQt6
PyQt6-SVG
cryptography
python-dotenv
argon2-cffi
requests